Version 1.2
With the help of David Fishburn, I was able to identify a problem and fix it. I also added a command and <Plug> mapping to activate the window without needing WinManager at all (for those who don't want to install WinManager). Thanks a lot to David for debugging the problem and providing help in other ways too.